Add check to build script for dist folder before trying to delete

+++ Extension/scripts/prod-build.js
@@ -5,7 +5,9 @@ const exec = util.promisify(require('child_process').exec);
 main();
 
 async function main() {
-  await fs.rm('./dist', { recursive: true });
+  if (fs.existsSync('./dist')) {
+    await fs.rm('./dist', { recursive: true });
+  }
   await fs.mkdir('./dist');
   await exec('npm run build -- --prod');
   console.log('Done');
